POO: Added evaluative challenges and last class challenge

+++ ads-section-3/programação-orientada-a-objeto/java/avaliativos/InverterNumero.java
@@ -4,13 +4,13 @@ public class InverterNumero {
     public static void main(String[] args) {
         Scanner read = new Scanner(System.in);
 
-        int numeroEntrada = read.nextInt(); // 123
-        float numeroFinal = 0;
+        int numeroEntrada = read.nextInt();
+        int numeroFinal = 0;
 
         while(numeroEntrada != 0) {
-            numeroFinal = numeroFinal * 10;                     // 0,3 * 10 = 3
-            numeroFinal = numeroFinal + (numeroEntrada % 10);   // 123 % 10 = 0,3
-            numeroEntrada = numeroEntrada / 10;                 // 123 / 10 = 12,3
+            numeroFinal = numeroFinal * 10;
+            numeroFinal = numeroFinal + (numeroEntrada % 10);
+            numeroEntrada = numeroEntrada / 10;
         }
 
         read.close();
